Test: getPostsByBookmark 관련 테스트 작성 파일 분리: PostsControllerTest -> GetPostsControllerTest 이름 변경: GetPostTest -> GetPostsServiceTest
